pyproject.toml
[project]
name = "ranking-metrics"
version = "0.1.0"
description = "A repository to understand ranking metrics as described by Musgrave et al. (2020)."
authors = [
    {name = "arrrrrmin", email = "armin.mueller@reutlingen-university.de"},
]
dependencies = [
    "torch>=2.0.1",
    "faiss-cpu>=1.7.4",
    "numpy>=1.25.2",
    "scikit-learn>=1.3.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.0",
    "icecream>=2.1.3",
    "torchshow>=0.5.1",
]
vis = [
    "matplotlib>=3.7.2",
]

[tool.black]
line-length = 100
